+++ changelog.md
@@ -4,12 +4,15 @@
 
 - **break**: make `src/build.task.ts`, `src/deploy.task.ts`,
   and `src/start.task.ts` work with SvelteKit
-  ([#157](https://github.com/feltcoop/gro/pull/157))
+  ([#157](https://github.com/feltcoop/gro/pull/157),
+  [#159](https://github.com/feltcoop/gro/pull/159))
   - add flag `gro deploy --clean` to reset deployment state
   - add flag `--branch` to both tasks, default to `main`
   - default to the `deploy` branch instead of `gh-pages`
 - **break**: rename `toEnvString` and `toEnvNumber` from `stringFromEnv` and `numberFromEnv`
   ([#158](https://github.com/feltcoop/gro/pull/158))
+- add helper `readDir` to `src/fs/node.ts`
+  [#159](https://github.com/feltcoop/gro/pull/159)
 
 ## 0.14.0